requirements.txt
setuptools>=68.0.0
fastapi==0.104.1
uvicorn==0.24.0
opencv-python==4.8.1.78
numpy==1.24.3
python-multipart==0.0.6
pydantic-settings==2.1.0
jinja2==3.1.2
aiofiles==23.2.1
mediapipe==0.10.9
mysql-connector-python
dlib
python-jose[cryptography]
starlette
itsdangerous
PyJWT